Release 1.1
This large patch (sorry, I wasn't able to make it into smaller chunks
this time) releases the 1.1 release. The main change is the unification
of Signal (object with listeners) and signal (descriptor) which in turn
makes it possible to have signals that are simple functions (not
classes).

There are a lot of documentation improvements and a lot more tests. In
retrospect, I think this is the first version that really works well.

Fixes: https://github.com/zyga/morris/issues/1
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>